requirements.txt
grpcio==1.37.0
grpcio-tools==1.37.0
pkg-resources==0.0.0
protobuf==3.15.8
six==1.15.0
